+++ src/app.js
@@ -24,3 +24,5 @@ app.use("/api-docs", swaggerUi.serve, swaggerUi.setup(swaggerDocument));
 
 module.exports = app;
 app.listen(3000,console.log("Server running on port 3000"));
+
+
